django_drf_react_live_chat_course/djchat/requirements.txt
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-csp==3.7
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
filelock==3.12.3
flake8==6.1.0
fsspec==2023.9.0
gradio_client==0.2.7
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
zope.interface==6.0
